refactor: 엔티티 구조 개선 및 할인 관련 함수 정리
