+++ src/refactoring/entities/coupon/index.ts
@@ -1,2 +1,3 @@
 export * from "./model";
 export * from "./types";
+export * from "./lib";
